+++ spec/support/tasks.rb
@@ -1,0 +1,9 @@
+require "rake"
+
+RSpec.configure do |config|
+  
+  config.before(:suite) do
+    Rails.application.load_tasks
+  end
+
+end